* src/org/jcae/mesh/Mesher.java   src/org/jcae/mesh/bora/algo/Basic2d.java: Do not call Mesh2D.removeDegeneratedEdges() before writing jcae2d files. All 2D algorithms can safely be applied to meshes loaded from disk.
* src/org/jcae/mesh/xmldata/MeshWriter.java (writeObjectNodes): 
Due to change just above, 2D nodes which represent the same
topological vertex are now all written onto disk.  But these
duplicates are written at the end of the file so that they
can be ignored when converting into 3D.

* src/org/jcae/mesh/xmldata/MeshToMMesh3DConvert.java
  src/org/jcae/mesh/bora/xmldata/Storage.java: When reading 2D
triangles, ignore triangles which are bound to a degenerated edge.
Duplicate nodes were already ignored.

* src/org/jcae/mesh/amibe/patch/Mesh2D.java (removeDegeneratedEdges):
Method removed, it is useless now.
* src/org/jcae/mesh/amibe/patch/VirtualHalfEdge2D.java (removeDegenerated):
Ditto.


git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@1984 9c4724a3-da1e-0410-bc68-88f18da35cd4