In pick(), throw an exception if color depth < 24 bits
In VTK, the following check is performed when entering
vtkVisibleCellSelector::Select() method:
  rwin->GetColorBufferSizes(rgba);
  if (rgba[0] < 8 || rgba[1] < 8 || rgba[2] < 8)
    {
    return;
    }

If color depth is lower than 24, picking fails silently,
which is quite unfortunate.  We throw an exception to
make user aware of the problem.

As GetColorBufferSizes method is not wrapped into Java,
we use GetDepthBufferSize instead.  This test may
give false negatives, so add a system property
org.jcae.vtk.checkColorDepth=false to bypass it.


git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@2479 9c4724a3-da1e-0410-bc68-88f18da35cd4