Make AbstractNode.parent final and do not call parent.addChild from constructors
A dirty hack was used to prevent childCreationListeners from being used
before initialized, a cleaner solution is to call parent.addChild
from LeafNode and Node constructors.

In Node constructor, childCreationListeners is still copied from its
parent so that it behaves like its parent when nodes are bound to it.
But childCreationListeners is no more reset when a Node is removed
from its parent.  This should have no impact since a Node is
normally dropped after being removed.

git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@2510 9c4724a3-da1e-0410-bc68-88f18da35cd4