Get rid of MappedByteBuffer
It is unlikely that
  http://bugs.sun.com/bugdatabase/view_bug.do;:YfiG?bug_id=4724038
gets fixed, and it is not clear how to use MappedByteBuffer
efficiently.

According to
  http://nadeausoftware.com/articles/2008/02/java_tip_how_read_files_quickly
MappedByteBuffer is not necessarily the fastest solution.
In order to perform tests, IntFileReader provides an interface
for reading file of ints, as with triangle connectivity.
It is implemented by IntFileReaderByMmap (which uses a MappedByteBuffer)
and IntFileReaderByDirectBuffer.
Initial tests perform on
  amibe/src/org/jcae/mesh/xmldata/tests/BenchmarkReadLargeFiles.java
show very similar results.

git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@2658 9c4724a3-da1e-0410-bc68-88f18da35cd4