Re-add a Metric2D member in Vertex2D
In SVN r2693, metrics have been removed from Vertex2D and
replaced by a map in Mesh2D.  But algos2d.Insertion takes
a lot of RAM because it creates many Vector2D instances,
and their metrics were kept in this map even if vertices
were eventually not added into the mesh.

A better caching strategy must be implemented so that
metrics of temporary vertices are not kept forever.
In the meantime, this map is removed and Metric2D member
of Vertex2D is restored.

git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@2696 9c4724a3-da1e-0410-bc68-88f18da35cd4