Fix several bugs in Remesh, there are no more inverted triangles when remeshing a cube
This is still not perfect, few sharp edges are swapped when edge
length is low.

Triangle centroids are not inserted for now, and vertices are not
projected onto local surface.

Add checkFindSurroundingTriangle() method in Remesh.java to check
sqrDistanceVertexTriangle().  This routine prints output files
which can be visualized by Medit.

git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@2838 9c4724a3-da1e-0410-bc68-88f18da35cd4