In Remesh(), make 2 passes to first discretize boundary and sharp edges
When remeshing a cube, few boundary edges were not preserved.
Here is a figure to show what happens.
         , A
        / \
       / M \
      /  *  \
   B +-------+ C
      \     /
       \   /
        \ /
         ' D
Vertex M is inserted by splitting edge [BC], and 4 triangles
are created: (ABM), (BDM), (DCM) and (CAM).
When [BC] is a sharp edge, original surface is not preserved,
and [BC] should not be split in that case.

But if [AB] or [AC] is split, insertion of vertex M may prevent
insertion of the midpoint of [BC], because this point would be too
near from M.

Current algorithm is:
 1  add all mesh nodes into an octree
 2  let NODES=empty set
 3  loop
 4    clear NODES
 5    for E in the set of mesh edges
 6      if (length(E) < sqrt(2))
 7        goto 5
 8      endif
 9      let M=midpoint(E)
10      if M not near of any other nodes of the octree
11        add M into octree
12        add M into NODES
13      endif
14    end
15    if NODES is empty
16      goto 24
17    endif
18    for M in NODES
19      let E=edge to split when inserting M
20      split E
21      locally improve triangle quality by swapping edges
22    end
23  end
24

To solve these problems, a first idea is to perform 2 passes:
 1. Discretize all boundary and sharp edges.
 2. Discretize inner edges, and do not split boundary and sharp edges.

Algorithm becomes:
 1  add all mesh nodes into an octree
 2  let NODES=empty set
 3  for pass in 0 1
 4    loop
 5      clear NODES
 6      for E in the set of mesh edges
 7        if pass == 0 and not (E is a sharp or boundary edge)
 8          goto 6
 9        endif
10        if pass == 1 and (E is a sharp or boundary edge)
11          goto 6
12        endif
13        if (length(E) < sqrt(2))
14          goto 6
15        endif
16        let M=midpoint(E)
17        if M not near of any other nodes of the octree
18          add M into octree
19          add M into NODES
20        endif
21      end
22      if NODES is empty
23        goto 3
24      endif
25      for M in NODES
26        let E=edge to split when inserting M
27        if pass == 1 and (E is a sharp or boundary edge)
28          goto 25
29        endif
30        split E
31        locally improve triangle quality by swapping edges
32      end
33    end
34  end
35

This change had been implemented, but if the new edge length is
much lower than original size, triangles after pass 0 are very flat,
and result mesh did contain degenerated triangles.

In the final algorithm, lines 3 and 4 are swapped.

git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@2842 9c4724a3-da1e-0410-bc68-88f18da35cd4