In Remesh, replace bgKdTree by neighborBgMap
When a mesh contains sharp edges, two faces may be very near
and kdTree.getNearestVertex() sometimes return a vertex on the
wrong side.  The resulting mesh is then badly broken.

bgKdTree is replaced by a map, each vertex of the current mesh
is bounded to a near vertex in the background mesh.

git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@2904 9c4724a3-da1e-0410-bc68-88f18da35cd4