Compute position of split points by dichotomy
Edge was previously discretized by a fixed number of segments,
and each point was tested to find a point at a distance 1 of
start point.  This was inaccurate, candidate point may be at
a distance >> 1 from start point, it could then be too near
from end point, and as a result no point was inserted on this
edge.

Now we use dichotomy to find a point at a distance 1 of start
point, this gives much accurate results and moreover requires
less computations.

git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@2964 9c4724a3-da1e-0410-bc68-88f18da35cd4