[vtk-util] Show/hide on leaf node did not work when they appear in the same manager node
When a manager node is displayed, Node.refreshData() is called; after leaves
are loaded, LeafNode.DataProvider.unLoad() is called to free memory, and
DataProvider does no more contain a geometry.  If a leaf node in the same
manager node is modified, this leaf node is thus no more visible.

DataProvider API is now modified to be more intuitive. load()
and unLoad() do nothing by default.

unLoad() may be overriden to calls the new DataProvider.clean() method
to free memory, and load() may be overriden to modified to reload geometry.