[Amibe] Allow caching projection when moving vertices
Vertices have to be projected onto the background mesh before
checking whether this vertex can be moved.  If not, it must be
reset to its previous location.
Traces should be temporarily disabled in order to not produce
inverted triangles before move is cancelled.
It is simpler and safer to store the previous location into a
cache, and print trace only when vertex is really moved.