[Amibe] Try again to fix findSurroundingTriangle()
As can be supposed from its name, findBetterTriangleInNeighborhood()
tries to find a better triangle in the neighborhood, but it does not
always return the best one.
The returned triangle is meant to be a better starting point for the
search algorithm in findSurroundingTriangle().  Thus make another
pass from this triangle.