[Amibe] Major rewrite, avoid calling MeshLiaison.findSurroundingTriangle()
We used to call MeshLiaison.findSurroundingTriangle() to retrieve
the triangle surrounding an inserted vertex, but it caused many problems
because it sometimes did not return the right triangle.

We could insert the vertex from within the outer loop on triangles, since
we then know the surrounding triangle.  Some care must be taken though,
vertex insertion adds new triangles, so triangle list must be copied
before iterating over it.

This commit uses a different approach.  We still compute the list of
candidate points beforehand, because this is efficient: no need to
copy triangle list, and all edges are discretized regularly.
Each triangle knows the list of candidates which belong to this triangle.
When a vertex is inserted, one triangle (for boundary edges), two triangles
(for interior manifold edges) or more (for non-manifold edges) are split.
For each triangle, those vertices are dispatched into adjacent triangles.
This is much more robust than the original scheme because only few triangles
have to be checked.