[Amibe] New Skeleton class
Vertices cannot always be inserted on boundary edges because vertices very
near of this edge may have been inserted before and prevents insertion
unless the allowNearnodes argument is passed.

This is in particular troublesome on cylindric patches, since they contain
flat triangles near boundaries.
We use here a new approach, feature edges of the mesh are stored in a
Skeleton class.  Before inserting a new vertex, we check whether it is
near an edge.  If yes, this vertex is moved on this edge.

This new approach removes the need of having different passes.