[Amibe] In buildRidges(), replace sharp edges by non-manifold edges
The buildRidges() method is now very similar to what buildGroupBoundaries().
This is not needed, recent commits already fixed bugs due to sharp edges,
but it seems wise to handle feature edges in a uniform way.

At the moment, edges are still tagged with SHARP attribute in order to
distinguish between ridges and other nonmanifold edges, but this may
change in a near future.